% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{addLocalFile}
\alias{addLocalFile}
\title{add vector data to leaflet map directly from the file system}
\usage{
addLocalFile(map, file, group = NULL)
}
\arguments{
\item{map}{a mapview or leaflet object.}

\item{file}{file path to the file to be added to \code{map}.}

\item{group}{the group name for the file to be added to \code{map}.}
}
\description{
add vector data to leaflet map directly from the file system
}
\examples{
library(leaflet)
library(sf)

destfile = tempfile(fileext = ".gpkg")

st_write(st_as_sf(gadmCHE), dsn = destfile)

leaflet() \%>\%
  addTiles() \%>\%
  leafem:::addLocalFile(destfile)

}
