% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtiles.R
\name{addPMTiles}
\alias{addPMTiles}
\title{Add vector tiles}
\usage{
addPMTiles(
  map,
  url,
  layerName,
  style = paintRules(layer = layerName),
  layerId = NULL,
  group = NULL
)
}
\arguments{
\item{map}{A \code{leaflet} or \code{mapview} map.}

\item{url}{The URL of vector tiles to add}

\item{layerName}{The name of the layer from the tiles to add}

\item{style}{Style elements to add}

\item{layerId}{The id of the layer to add}

\item{group}{The group that the layer belongs to if any}
}
\description{
This function add vector tiles, as created by
[tippecanoe](https://github.com/protomaps/tippecanoe), a software project
that builds vector tiles stored as .mbtiles files originally developed by
MapBox but subsequently maintained by protomaps.
}
\examples{
library(leaflet)
u_mbtiles = "https://www.carbon.place/tiles/pct/{x}/{y}/{z}.pbf"
leaflet() \%>\% addProviderTiles("OpenStreetMap") \%>\% addPMTiles(u_mbtiles, layerName = "bicycle")
}
