% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRasterRGB.R
\name{addRasterRGB}
\alias{addRasterRGB}
\alias{addStarsRGB}
\title{Add an RGB image as a layer}
\usage{
addRasterRGB(map, x, r = 3, g = 2, b = 1, quantiles = c(0.02,
  0.98), maxpixels = 5e+05, na.color = "#BEBEBE80",
  method = c("bilinear", "ngb"), ...)

addStarsRGB(map, x, r = 3, g = 2, b = 1, quantiles = c(0.02, 0.98),
  maxpixels = 5e+05, na.color = "#BEBEBE80", method = c("bilinear",
  "ngb"), ...)
}
\arguments{
\item{map}{a map widget object created from `leaflet()``}

\item{x}{a  `RasterBrick`, `RasterStack` or `stars`` raster object}

\item{r}{integer. Index of the Red channel/band, between 1 and nlayers(x)}

\item{g}{integer. Index of the Green channel/band, between 1 and nlayers(x)}

\item{b}{integer. Index of the Blue channel/band, between 1 and nlayers(x)}

\item{quantiles}{the upper and lower quantiles used for color stretching.
If set to NULL, no stretching is applied.}

\item{maxpixels}{integer > 0. Maximum number of cells to use for the plot.
If maxpixels < \code{ncell(x)}, sampleRegular is used before plotting.}

\item{na.color}{the color to be used for NA pixels}

\item{method}{Method used to compute
values for the resampled layer that is passed on to leaflet. mapview does
projection on-the-fly to ensure correct display and therefore needs to know
how to do this projection. The default is 'bilinear' (bilinear interpolation),
which is appropriate for continuous variables. The other option, 'ngb'
(nearest neighbor), is useful for categorical variables.}

\item{...}{additional arguments passed on to \code{\link{addRasterImage}}}
}
\description{
Create a Red-Green-Blue image overlay from a \code{RasterStack} /
\code{RasterBrick} or \code{stars} object based on three layers.
Three layers (sometimes referred to as "bands" because they may represent
different bandwidths in the electromagnetic spectrum) are combined such
that they represent the red, green and blue channel. This function can
be used to make 'true (or false) color images' from Landsat and other
multi-band satellite images. Note, this text is plagirized, i.e. copied
from \code{\link{plotRGB}}.
\code{AddRasterRGB} and \code{addStarsRGB} are aliases.
}
\examples{
if (interactive()) {
  library(raster)
  library(stars)
  library(plainview)
  library(leaflet)

  leaflet() \%>\%
    addTiles(group = "OpenStreetMap") \%>\%
    addRasterRGB(plainview::poppendorf, 4,3,2, group = "True colours") \%>\%
    addStarsRGB(st_as_stars(plainview::poppendorf), 5,4,3, group = "False colours") \%>\%
    addLayersControl(
      baseGroups = c("Satellite"),
      overlayGroups = c("True colours", "False colours"),
    )
}

}
\author{
Tim Appelhans, Luigi Ranghetti
}
