% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{latticeView}
\alias{latticeView}
\alias{latticeview}
\alias{sync}
\title{View two or more (possibly synchronised) mapview or leaflet maps}
\usage{
latticeView(..., ncol = 2, sync = "none", sync.cursor = FALSE,
  no.initial.sync = TRUE)

latticeview(...)

sync(..., ncol = 2, sync = "all", sync.cursor = TRUE,
  no.initial.sync = TRUE)
}
\arguments{
\item{...}{any number of mapview or leaflet objects or a list thereof}

\item{ncol}{how many columns should be plotted}

\item{sync}{whether to synchronise zoom and pan for certain elements.
Possible values are "all" (default) to sync all maps, "none" to disable
synchronisation or a list of panel numbers, e.g. \code{list(c(1, 3), c(2, 4))}
will synchronise panels 1 & 3 and panels 2 & 4. Panels are drawn from top right
to bottom left.}

\item{sync.cursor}{whether to show cursor position in synced panels (default TRUE).}

\item{no.initial.sync}{whether to sync the initial view (default TRUE).}
}
\description{
This function produces a lattice like view of two or more maps. It is possible to sync
any combination of panels or all or none. For synchronising all panels it
is best to use the provided convenience function \code{sync}.
}
\section{Functions}{
\itemize{
\item \code{latticeview}: alias for ease of typing

\item \code{sync}: convenience function for syncing maps
}}

\examples{
\dontrun{
library(leaflet)
library(leafsync)

m1 = leaflet() \%>\%
  addTiles() \%>\%
  addPolygons(data = gadmCHE)

m2 = leaflet() \%>\%
  addProviderTiles("Esri.WorldImagery") \%>\%
  addCircleMarkers(data = breweries91,
                   color = "black",
                   opacity = 0.9,
                   fillColor = "white",
                   fillOpacity = 0.9)

# synced
sync(m1, m2)
sync(m1, m2, no.initial.sync = TRUE)

# not synced
latticeview(m1, m2)
latticeview(m1, m2, ncol = 1)
}

}
