% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeColors.R
\name{changeColors}
\alias{changeColors}
\title{Change the color palette of a map layer}
\usage{
changeColors(map, className, colors)
}
\arguments{
\item{map}{a mapview or leaflet object.}

\item{className}{character; the class name to apply the color-change to. The layer(s)
must have had this class name assigned to it; see examples}

\item{colors}{character vector; the colors that form the new color palette. Colors dd
can be either named colors in R (like "red" or "blue") or hexadecimal colors}
}
\description{
Given a class name that corresponds to a map layer or layers, uses the
'gradientmap' JavaScript library to change the color scheme on the fly
}
\examples{
if (interactive()) {
  library(leafem)
  library(leaflet)

  leaflet() \%>\%
    addTiles() \%>\%
    addWMSTiles("https://www.mrlc.gov/geoserver/mrlc_display/NLCD_2016_Bare_Ground_Shrubland_Fractional_Component/ows?SERVICE=WMS&",
                layers = "NLCD_2016_Bare_Ground_Shrubland_Fractional_Component",
                options = WMSTileOptions(className = "bare_ground", transparent = TRUE, format = "image/png")) \%>\%
    changeColors("bare_ground", terrain.colors(20))

  leaflet() \%>\%
    addTiles(options = tileOptions(className = "base")) \%>\%
    changeColors("base", colorRampPalette(c("red", "white"))(50))
}
}
