% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeoRaster.R
\name{imagequeryOptions}
\alias{imagequeryOptions}
\title{Imagequery options for addGeoRaster, addGeotiff and addCOG}
\usage{
imagequeryOptions(
  className = "info legend",
  position = c("topright", "topleft", "bottomleft", "bottomright"),
  type = c("mousemove", "click"),
  digits = NULL,
  prefix = "Layer",
  noData = "NoData Value"
)
}
\arguments{
\item{className}{a character string to append to the control legend.}

\item{position}{where to place the display field. Default is 'topright'.}

\item{type}{whether query should occur on 'mousemove' or 'click'. Defaults
to 'mousemove'.}

\item{digits}{the number of digits to be shown in the display field.}

\item{prefix}{a character string to be shown as prefix for the layerId.}

\item{noData}{the text shown when the mouse is over a `NoData Value` as
defined by GDAL. The default "NoData Value" will use whatever is defined in the
Geotif.}
}
\description{
Imagequery options for addGeoRaster, addGeotiff and addCOG
}
