% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtiles.R
\name{paintRules}
\alias{paintRules}
\title{Styling options for PMTiles}
\usage{
paintRules(
  layer,
  fillColor = "#0033ff66",
  color = "#0033ffcc",
  do_stroke = TRUE,
  width = 0.5,
  radius = 3,
  stroke = "#000000",
  opacity = 1
)
}
\arguments{
\item{fillColor}{fill color for polygons}

\item{color}{line color}

\item{do_stroke}{logical, whether polygon borders should be drawn}

\item{width}{line width}

\item{radius}{point radius}

\item{stroke}{color point border}

\item{opacity}{point opacity}
}
\description{
Styling options for PMTiles
}
