% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtiles.R
\name{paintRules}
\alias{paintRules}
\title{Styling options for PMTiles}
\usage{
paintRules(
  layer,
  fillColor = "#0033ff66",
  color = "#0033ffcc",
  do_stroke = TRUE,
  width = 0.5,
  radius = 3,
  stroke = "#000000",
  opacity = 1,
  dash = NULL
)
}
\arguments{
\item{layer}{the name of the layer in the PMTiles file to visualise.}

\item{fillColor}{fill color for polygons}

\item{color}{line color}

\item{do_stroke}{logical, whether polygon borders should be drawn}

\item{width}{line width}

\item{radius}{point radius}

\item{stroke}{color point border}

\item{opacity}{point opacity}

\item{dash}{either `NULL` (default) for a solid line or a numeric vector
of length 2 denoting segment length and spce between segments (in pixels),
e.g. `c(5, 3)`}
}
\description{
Styling options for PMTiles
}
